+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Gandi V5 API Gem Changelog
 
+## Version 0.9.0
+
 * Add transferring a domain to Gandi (I don't have any domains outside Gandi to test this myself so it's possible I've misread the docs and a bug is waiting to be found, please add an issue if I have):
   * GandiV5::Domain::TransferIn:
     * .create(fqdn, \*\*options)